Initial Root segmentation Coded
The code is now setup to use the correct channels, and can take an image and attempts to segment the roots according to the correct model file and correct channels.

Output is currently different than if it were ran in the Color Clustering program itself, so there are some settings not quite right as of now. Committing because it's still a functional change.