added ChunkedOverlay
Added a method to overlay the chunks of the segmentation phase. I am fairly certain the rectangles overlayed are not actually the chunks from the segmentation phase, but I am not sure how to properly unpack the chunks.

Added that method to help with debugging the paths problem. Still not progress on that.